Hide risky settings
Added a new setting Enable Risky Features, defaults to False.
Hides Enable Ticket Rate Loss Logging if False and prevents
it from being set to True.

If installed into a config that already has it set to True,
it will be forced to False with a warning.

The only way to Enable Ticket Rate Loss Logging is to
Enable Risky Features first, but that setting is hidden.

The only way to show that setting is to use the undocumented
command "risky show". It can be hidden again with "risky hide".

The idea is that everyone that has it set to True before
the update will have it forced to False.

Only testers who are given the "risky show" undocumented
command will be able to enable it again.